Refactor: Move README-specific information out of  `ClassInterface`.
Remove unneed comments.
