+++ code/drasil-code/lib/Language/Drasil/Code/Imperative/GOOL/ClassInterface.hs
@@ -3,7 +3,6 @@
 -- | Defines a package extension for GOOL, with functions for pairing a GOOL
 -- program with auxiliary, non-source-code files.
 module Language.Drasil.Code.Imperative.GOOL.ClassInterface (
-  ReadMeInfo(..),
   -- Typeclasses
   PackageSym(..), AuxiliarySym(..)
 ) where
@@ -11,8 +10,8 @@ module Language.Drasil.Code.Imperative.GOOL.ClassInterface (
 import Language.Drasil (Expr)
 import Database.Drasil (ChunkDB)
 import Language.Drasil.Code.DataDesc (DataDesc)
-import Language.Drasil.Mod (Name, Version)
 import Language.Drasil.Choices (Comments, ImplementationType, Verbosity)
+import Language.Drasil.Code.Imperative.ReadMe.Import (ReadMeInfo(..))
 
 import Drasil.GOOL (ProgData, GOOLState)
 
@@ -41,42 +40,3 @@ class AuxiliarySym r where
 
   auxHelperDoc :: r (AuxHelper r) -> Doc
   auxFromData :: FilePath -> Doc -> r (Auxiliary r)
-
--- | Language name.
-type LangAbbrev = String
--- | Programming language version.
-type LangVers = String
--- | Case name.
-type CaseName = String
--- | Purpose of example
-type ExamplePurpose = String
--- | Description of example
-type ExampleDescr = String
--- | Motivation of example
-type ExampleMotivation = String
--- | Scope of exmample
-type ExampleScope = String
--- | File contributors
-type Contributor = String
--- | Input File
-type InFile = String -- TODO: There may not always be an Input/Output File
--- | Output File
-type OutFile = String
--- | Holds all information needed to create a README file.
-data ReadMeInfo = ReadMeInfo {
-  langName :: LangAbbrev,
-  langVersion :: LangVers,
-  invalidOS :: Maybe String,
-  implementType :: ImplementationType,
-  extLibNV :: [(Name,Version)],
-  extLibFP :: [FilePath],
-  contributors :: [Contributor],
-  configFP :: [FilePath],
-  caseName :: CaseName,
-  examplePurpose :: ExamplePurpose,
-  exampleDescr :: ExampleDescr,
-  exampleMotivation :: ExampleMotivation,
-  exampleScope :: ExampleScope,
-  folderNum :: Int,
-  inputOutput :: (InFile, OutFile)
-}
